Moved components to kraken_components directory.
