+++ Python/kraken_components/biped/clavicle_component.py
