+++ Python/kraken_components/biped/foot_component.py
