+++ Python/kraken_components/biped/hand_component.py
@@ -151,7 +151,6 @@ class HandComponentGuide(HandComponent):
         self.digitNamesAttr.setValue(data.get('digitNames'))
 
         fingersGuideXfos = data.get('fingersGuideXfos')
-        fingerShapeCtrlData = data.get('fingerShapeCtrlData')
         if fingersGuideXfos is not None:
 
             for finger in self.fingers.keys():
@@ -159,7 +158,9 @@ class HandComponentGuide(HandComponent):
                     self.fingers[finger][i].xfo = fingersGuideXfos[finger][i]
 
                     if hasattr(self.fingers[finger][i], 'shapeCtrl'):
-                        self.fingers[finger][i].shapeCtrl.setCurveData(fingerShapeCtrlData[finger][i])
+                        fingerShapeCtrlData = data.get('fingerShapeCtrlData')
+                        if fingerShapeCtrlData is not None:
+                            self.fingers[finger][i].shapeCtrl.setCurveData(fingerShapeCtrlData[finger][i])
 
         return True
