+++ Python/kraken_components/biped/leg_component.py
