+++ Python/kraken_components/fabrice/fabrice_head.py
