+++ Python/kraken_components/fabrice/fabrice_rig.krg
