+++ Python/kraken_components/fabrice/fabrice_tail.py
