+++ Python/kraken_components/generic/insectleg_component.py
