+++ Python/kraken_components/generic/mainSrt_component.py
