+++ Python/kraken_components/generic/tentacle_component.py
