+++ Python/kraken_examples/spineClav_rig.py
@@ -3,8 +3,8 @@ from kraken.core.maths import Vec3, Quat, Xfo
 from kraken.core.objects.container import Container
 from kraken.core.objects.layer import Layer
 
-from kraken_examples.clavicle_component import ClavicleComponentGuide, ClavicleComponentRig
-from kraken_examples.spine_component import SpineComponentRig
+from kraken_components.biped.clavicle_component import ClavicleComponentGuide, ClavicleComponentRig
+from kraken_components.biped.spine_component import SpineComponentRig
 
 from kraken.core.profiler import Profiler
 from kraken.helpers.utility_methods import logHierarchy
